+++ apps/web/src/features/nodes/components/NodesGeneric.tsx
@@ -3,13 +3,13 @@
 import React, { Suspense } from 'react';
 import { NodeProps } from '@xyflow/react';
 import { GenericNode } from './base/GenericNode';
-import { UNIFIED_NODE_TYPES } from '@/shared/types';
+import { UNIFIED_NODE_CONFIGS } from '@/shared/types';
 import { StartNode, ConditionNode, JobNode, DBNode, EndpointNode, PersonJobNode, PersonBatchJobNode } from './nodes';
 
 // Loading component for lazy-loaded nodes
 const NodeLoadingFallback: React.FC<NodeProps> = ({ id, data, selected }) => {
   const nodeType = (data as { type: string }).type;
-  const config = UNIFIED_NODE_TYPES[nodeType];
+  const config = UNIFIED_NODE_CONFIGS[nodeType];
   const reactFlowType = config?.reactFlowType || "startNode";
   
   return (
@@ -36,7 +36,7 @@ const UniversalNode: React.FC<NodeProps> = (props) => {
   const NodeComponent = nodeComponents[nodeType];
   
   if (!NodeComponent) {
-    const config = UNIFIED_NODE_TYPES[nodeType];
+    const config = UNIFIED_NODE_CONFIGS[nodeType];
     return (
       <GenericNode {...props} nodeType={config?.reactFlowType || "startNode"}>
         <span className="text-red-500">Unknown node type: {nodeType}</span>
@@ -52,9 +52,9 @@ const UniversalNode: React.FC<NodeProps> = (props) => {
 };
 
 
-// Default export mapping for React Flow using UNIFIED_NODE_TYPES
+// Default export mapping for React Flow using UNIFIED_NODE_CONFIGS
 export default Object.fromEntries(
-  Object.entries(UNIFIED_NODE_TYPES).map(([_, config]) => [
+  Object.entries(UNIFIED_NODE_CONFIGS).map(([_, config]) => [
     config.reactFlowType,
     UniversalNode
   ])
