+++ apps/web/src/features/nodes/components/nodes/EndpointNode.tsx
@@ -2,10 +2,10 @@ import React from 'react';
 import { NodeProps } from '@xyflow/react';
 import { GenericNode } from '../base/GenericNode';
 import { Save } from 'lucide-react';
-import { UNIFIED_NODE_TYPES, type EndpointBlockData } from '@/shared/types';
+import { UNIFIED_NODE_CONFIGS, type EndpointBlockData } from '@/shared/types';
 
 const EndpointNode: React.FC<NodeProps> = ({ id, data, selected }) => {
-  const config = UNIFIED_NODE_TYPES.endpoint;
+  const config = UNIFIED_NODE_CONFIGS.endpoint;
   const endpointData = data as EndpointBlockData;
   
   return (
