+++ apps/web/src/features/nodes/components/nodes/StartNode.tsx
@@ -1,10 +1,10 @@
 import React from 'react';
 import { NodeProps } from '@xyflow/react';
 import { GenericNode } from '../base/GenericNode';
-import { UNIFIED_NODE_TYPES } from '@/shared/types';
+import { UNIFIED_NODE_CONFIGS } from '@/shared/types';
 
 const StartNode: React.FC<NodeProps> = ({ id, data, selected }) => {
-  const config = UNIFIED_NODE_TYPES.start;
+  const config = UNIFIED_NODE_CONFIGS.start;
   
   return (
     <GenericNode id={id} data={data} selected={selected} nodeType={config.reactFlowType}>
