+++ apps/web/src/features/properties/components/PropertiesRenderer.tsx
@@ -1,7 +1,7 @@
 // Reusable component for rendering property panels based on selection
 import React from 'react';
 import { UniversalPropertiesPanel } from '@/features/properties';
-import { DiagramNode, Arrow, PersonDefinition, ArrowData } from '@/shared/types/domain';
+import { DiagramNode, Arrow, PersonDefinition, ArrowData } from '@/shared/types/core';
 
 interface PropertiesRendererProps {
   selectedNodeId?: string | null;
