+++ apps/web/src/features/properties/components/UniversalPropertiesPanel.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Settings } from 'lucide-react';
 import { Panel } from '../wrappers';
-import { UNIFIED_NODE_TYPES } from '@/shared/types';
+import { UNIFIED_NODE_CONFIGS } from '@/shared/types';
 import { PanelConfig } from '@/shared/types/panelConfig';
 import { GenericPropertyPanel } from './ui-components/GenericPropertyPanel';
 import {
@@ -32,7 +32,7 @@ const PANEL_CONFIGS: Record<string, PanelConfig<any>> = {
 export const UniversalPropertiesPanel: React.FC<{ nodeId: string; data: any }> = ({ nodeId, data }) => {
   const nodeType = data.type;
   const panelConfig = PANEL_CONFIGS[nodeType];
-  const nodeConfig = UNIFIED_NODE_TYPES[nodeType];
+  const nodeConfig = UNIFIED_NODE_CONFIGS[nodeType];
   
   if (!panelConfig) {
     return (
