--- apps/web/src/shared/types/domain.ts
@@ -1,154 +1,0 @@
-// --- File: apps/web/src/shared/types/types.ts ---
-// (Expanded and corrected)
-import {Node, Edge, OnEdgesChange, EdgeChange, Connection,
-  applyEdgeChanges as applyEdgeChangesRF,
-  addEdge as addEdgeRF
-} from '@xyflow/react';
-
-
-export interface ApiKey {
-  id: string;
-  name: string;
-  service: 'claude' | 'chatgpt' | 'grok' | 'gemini' | 'custom';
-  // Key itself should NOT be stored in frontend state directly if possible,
-  // rather managed by a secure backend or encrypted in browser storage.
-  keyReference?: string;
-}
-
-export type ArrowKind = 'normal' | 'fixed';
-export type BlockType = 'start' | 'person_job' | 'person_batch_job' | 'db' | 'job' | 'condition' | 'endpoint';
-
-// Base for all canvas blocks (nodes) with discriminating `type`
-export interface BaseBlockData extends Record<string, unknown> {
-  id: string;
-  type: BlockType;
-  label: string;
-  flipped?: boolean;
-}
-
-export interface StartBlockData extends BaseBlockData {
-  type: 'start';
-}
-
-export interface PersonJobBlockData extends BaseBlockData {
-  type: 'person_job';
-  personId?: string;
-  llmApi?: ApiKey['service'];
-  apiKeyId?: ApiKey['id'];
-  modelName?: string;
-  defaultPrompt?: string;
-  firstOnlyPrompt?: string;
-  detectedVariables?: string[];
-  contextCleaningRule?: 'uponRequest' | 'noForget' | 'onEveryTurn';
-  contextCleaningTurns?: number;
-  iterationCount?: number;
-}
-
-export interface PersonBatchJobBlockData extends BaseBlockData {
-  type: 'person_batch_job';
-  personId?: string;
-  llmApi?: ApiKey['service'];
-  apiKeyId?: ApiKey['id'];
-  modelName?: string;
-  batchPrompt?: string;
-  batchSize?: number;
-  parallelProcessing?: boolean;
-  aggregationMethod?: 'concatenate' | 'summarize' | 'custom';
-  customAggregationPrompt?: string;
-  detectedVariables?: string[];
-  iterationCount?: number;
-}
-
-export type JobBlockSubType = 'api_tool' | 'diagram_link' | 'code';
-export interface JobBlockData extends BaseBlockData {
-  type: 'job';
-  subType: JobBlockSubType;
-  sourceDetails: string;
-}
-
-export type DBBlockSubType = 'fixed_prompt' | 'file';
-
-
-export interface DBBlockData extends BaseBlockData {
-  type: 'db';
-  subType: DBBlockSubType;
-  sourceDetails: string;
-}
-
-
-export type ConditionType = 'expression' | 'max_iterations';
-export interface ConditionBlockData extends BaseBlockData {
-  type: 'condition';
-  conditionType: ConditionType;
-  expression?: string;
-  maxIterations?: number;
-}
-
-
-export interface EndpointBlockData extends BaseBlockData {
-  type: 'endpoint';
-  // Optional file save properties
-  saveToFile?: boolean;
-  filePath?: string;
-  fileFormat?: 'json' | 'text' | 'csv';
-}
-
-// Person definitions live outside the canvas; used to configure LLM settings
-export interface PersonDefinition {
-  id: string;
-  label: string;
-  service?: ApiKey['service'];
-  apiKeyId?: ApiKey['id'];
-  modelName?: string;
-  systemPrompt?: string;
-}
-
-export interface ArrowData extends Record<string, unknown> {
-  id: string;
-  sourceBlockId: string;
-  targetBlockId: string;
-  sourceHandleId?: string;
-  targetHandleId?: string;
-  label?: string;
-  contentType?: 'raw_text' | 'variable_in_object' | 'conversation_state';
-  arrowKind?: ArrowKind;
-  variableName?: string;
-  objectKeyPath?: string;
-  // iterationBehavior?: 'once' | 'each';
-  loopRadius?: number;
-  branch?: 'true' | 'false';
-  // Control point offset from default midpoint (for custom arrow curves)
-  controlPointOffsetX?: number;
-  controlPointOffsetY?: number;
-}
-
-// Union type for all block data types  
-export type DiagramNodeData = StartBlockData | PersonJobBlockData | PersonBatchJobBlockData | JobBlockData | DBBlockData | ConditionBlockData | EndpointBlockData;
-
-// Type for diagram nodes with React Flow compatibility
-export type DiagramNode = Node<DiagramNodeData & Record<string, unknown>>;
-
-export interface DiagramState {
-  // Person definitions shown in top bar, not as nodes
-  persons: PersonDefinition[];
-  nodes: DiagramNode[];
-  arrows: Edge<ArrowData>[];
-  apiKeys: ApiKey[];
-}
-
-export function applyArrowChanges(
-  changes: EdgeChange[],
-  arrows: Arrow<ArrowData>[]
-): Arrow<ArrowData>[] {
-  return applyEdgeChangesRF(changes, arrows) as Arrow<ArrowData>[];
-}
-export function addArrow(
-  arrow: Edge<ArrowData> | Connection,
-  arrows: Edge<ArrowData>[]
-): Edge<ArrowData>[] {
-  return addEdgeRF(arrow, arrows);
-}
-
-export type ArrowChange = EdgeChange;
-export type Arrow<T extends Record<string, unknown> = ArrowData> = Edge<T>;
-export type OnArrowsChange = OnEdgesChange;
