+++ apps/web/src/shared/types/index.ts
@@ -1,6 +1,5 @@
-export * from './domain'
-export * from './nodeType'
+export * from './core'
+export * from './nodeConfig'
 export * from './errorHandling'
 export * from './components'
 export * from './properties'
-export * from './execution'
